Add files via upload
This report was build for CEO office to directly get details of Bids/ Proposal sent and what's the status of that bid.
Bids are bifurcated based on timeline i.e. , when it is due and what state account for how much bud value and who is sponsor of that project.